% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_sim.R
\name{corr_sim}
\alias{corr_sim}
\title{Simulate a multivariate covariance/correlation matrix}
\usage{
corr_sim(
  n_units = 10,
  n_spp = 3,
  spp_mu = c(0, 0, 0),
  error_name = "spp_error_psi",
  sigma_in = NULL
)
}
\arguments{
\item{n_units}{number of sampling units to simulate}

\item{n_spp}{number of species to simulate}

\item{spp_mu}{means for value}

\item{error_name}{name for MVN error}

\item{sigma_in}{MVN sigma, default is N}
}
\value{
Returns a list with the following a list containing:
\item{sigma_sim}{The simulated errors in matrix format.}
\item{unit_obs_spp}{The simulated errors in tibble format.}
\item{sigma}{The input value used for \code{sigma}}
}
\description{
The \code{sigma} input allows you specify the default covariance. Leave
1's along the diagonal to simulate according to a covariance matrix that is
equivalent to its correlation matrix. Alternatively,
you use random values, which is the default, by leaving this as \code{NULL}.
}
\details{
The default is a three species site occupancy correlation matrix with 10
simulates units.

The output include two formats, one as the raw matrix for use with functions
like \code{cor()} to examine the outputs and a second as tibble for use when
simulating data.
}
\examples{
# example code
example_out <-
  corr_sim(n_units = 10,
           n_spp = 3,
           spp_mu = c(0, 0, 0),
           error_name = "spp_error_psi",
           sigma_in = NULL)
cor(example_out$sigma_sim)
}
