% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occstanhm-package.R
\docType{package}
\name{occstanhm-package}
\alias{occstanhm}
\alias{occstanhm-package}
\title{The 'occstanhm' package.}
\description{
Methods for \emph{occ}-upancy models in \emph{Stan} using \emph{h}-ierarhcial
\emph{m}-ethods (hence \code{occstanhm}). The packages is specifically designed for
fitting multispecies, hierarchical Bayesian occupancy models using Stan,
created with eDNA-based models in mind.
Currently, this includes a two-level and three-level models.
}
\author{
\strong{Maintainer}: Richard Erickson \email{rerickson@usgs.gov} (\href{https://orcid.org/0000-0003-4649-482X}{ORCID})

Authors:
\itemize{
  \item Charlie Labuzzetta \email{clabuzzetta@usgs.gov} (\href{https://orcid.org/0000-0002-6027-0120}{ORCID})
}

}
\keyword{internal}
