% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occstanhm_fast_2.R
\name{occstanhm_fast_2}
\alias{occstanhm_fast_2}
\title{Two-level, multispecies hierarchical occupancy model}
\usage{
occstanhm_fast_2(
  stan_data,
  n_chains = 2,
  n_parallel_chains = 2,
  n_threads_per_chain = 4,
  n_refresh = 100,
  n_warmup = 200,
  n_sample = 200,
  ...
)
}
\arguments{
\item{stan_data}{Data formatted for Stan.}
}
\value{
fit Model fit from cmdstanr.
}
\description{
occstanhm_2 is a two-level, hierarchical occupancy model.
occstanhm_fast_2 is a stripped down version of the two-level model designed
to be comparable to Tolber et al. 2019 (https://doi.org/10.1002/ecy.2754).
}
\details{
The model does not included hyper-parameters and a correlation matrix for the
p-level fo the model. The model also allows for a single detection
probability to be estimated across all species.

For an example, see the comarpison vignette.
}
\examples{

\dontrun{

}
}
