% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_long_form_occ_2.R
\name{sim_long_form_occ_2}
\alias{sim_long_form_occ_2}
\title{Simulate long-form 2-level occupancy data data}
\usage{
sim_long_form_occ_2(
  n_units = 50,
  n_unit_revisit_mean = 6,
  n_spp = 4,
  k_samples_mean = 10,
  spp_mu = rep(0.5, n_spp),
  sigma_psi = matrix(c(1, 0, 0, 0, 0, 1, 1, -0.7, 0, 1, 1, -0.7, 0, -0.7, -0.7, 1), nrow
    = n_spp, ncol = n_spp),
  sigma_p = matrix(c(1, 0, 0, 0, 0, 1, 1, -0.7, 0, 1, 1, -0.7, 0, -0.7, -0.7, 1), nrow =
    n_spp, ncol = n_spp),
  psi_predictor = NULL,
  p_predictor = NULL,
  p_same = FALSE
)
}
\arguments{
\item{n_units}{Number of units to simulate}

\item{n_unit_revisit_mean}{Poisson average for revisits per unit}

\item{n_spp}{number of species}

\item{k_samples_mean}{Poisson average number of samples per revisit}

\item{spp_mu}{logit-scale average for p and psi for each species}

\item{sigma_psi}{variance/co-variance matrix for psi}

\item{sigma_p}{variance/co-variance matrix for p}

\item{psi_predictor}{prediction row (e.g., row sum of design matrix)}

\item{p_predictor}{prediction row (e.g., row sum of design matrix)}

\item{p_same}{Boolean if det prob p is the same across all species and units}
}
\value{
Returns a list with the following a list containing:
\item{sigma_corr_psi}{The simulted error outputs used for psi}
\item{sigma_corr_p}{The simulted error outputs used for p}
\item{spp_obs}{A simulated dataset.}
}
\description{
This code simulates long-form (in contrast to wide, matrix format) data. The
function is used both to demonstrate and test the package. The function
simulates random numbers of revisits and samples to create more realistic
data, but also limit confidences that occurs when using fixed numbers for
these values.
}
\examples{
example_sim <-
  sim_long_form_occ_2()
}
