% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_long_form_occ_3.R
\name{sim_long_form_occ_3}
\alias{sim_long_form_occ_3}
\title{Simulate long form data for a 3-level occupancy model}
\usage{
sim_long_form_occ_3(
  n_units = 50,
  n_unit_revisit_mean = 6,
  k_samples_mean = 10,
  k_subsamples = 4,
  n_spp = 4,
  spp_mu = seq(-0.5, 1, length.out = n_spp),
  sigma_psi = NULL,
  sigma_theta = NULL,
  sigma_p = NULL
)
}
\arguments{
\item{n_units}{Number of units to simulate}

\item{n_unit_revisit_mean}{Poisson average for revisits per unit}

\item{k_samples_mean}{Poisson average number of samples per revisit}

\item{k_subsamples}{number of subsamples per sample (molecular replicates)}

\item{n_spp}{number of species}

\item{spp_mu}{logit-scale average for p and psi for each species}

\item{sigma_psi}{variance/co-variance matrix for psi}

\item{sigma_theta}{variance/co-variance matrix for theta}

\item{sigma_p}{variance/co-variance matrix for p}
}
\value{
Returns a list with the following a list containing:
\item{sigma_corr_psi}{The simulted error outputs used for psi}
\item{sigma_corr_p}{The simulted error outputs used for p}
\item{spp_obs}{A simulated dataset.}
}
\description{
Simulates data in a long form for a 3-level occupancy model
}
